Rename threads.py to threading.py
Structured log message prefix is based on module path,
so renaming threads to threading should improve log message readability.
