+++ pyanaconda/install.py
@@ -30,7 +30,7 @@ from pyanaconda import timezone
 from pyanaconda import network
 from pyanaconda import screen_access
 from pyanaconda.i18n import N_
-from pyanaconda.threads import threadMgr
+from pyanaconda.threading import threadMgr
 from pyanaconda.ui.lib.entropy import wait_for_entropy
 from pyanaconda.kickstart import runPostScripts, runPreInstallScripts
 from pyanaconda.kexec import setup_kexec
