+++ pyanaconda/ntp.py
@@ -29,7 +29,7 @@ import ntplib
 import socket
 
 from pyanaconda import isys
-from pyanaconda.threads import threadMgr, AnacondaThread
+from pyanaconda.threading import threadMgr, AnacondaThread
 from pyanaconda.constants import THREAD_SYNC_TIME_BASENAME
 
 NTP_CONFIG_FILE = "/etc/chrony.conf"
