+++ pyanaconda/rescue.py
@@ -23,7 +23,7 @@ from blivet.osinstall import mount_existing_system, find_existing_installations
 from pyanaconda import iutil
 from pyanaconda.constants import ANACONDA_CLEANUP, THREAD_STORAGE
 from pyanaconda.constants_text import INPUT_PROCESSED
-from pyanaconda.threads import threadMgr
+from pyanaconda.threading import threadMgr
 from pyanaconda.flags import flags
 from pyanaconda.i18n import _, N_, C_
 from pyanaconda.kickstart import runPostScripts
