+++ pyanaconda/ui/gui/spokes/advstorage/dasd.py
@@ -26,7 +26,7 @@ from pyanaconda.ui.gui import GUIObject
 from pyanaconda.ui.gui.utils import gtk_action_nowait
 from pyanaconda.storage_utils import try_populate_devicetree
 from pyanaconda.regexes import DASD_DEVICE_NUMBER
-from pyanaconda.threads import threadMgr, AnacondaThread
+from pyanaconda.threading import threadMgr, AnacondaThread
 from pyanaconda import constants
 
 __all__ = ["DASDDialog"]
