+++ pyanaconda/ui/gui/spokes/software.py
@@ -26,7 +26,7 @@ from gi.repository import Gtk, Pango
 from pyanaconda.flags import flags
 from pyanaconda.i18n import _, C_, CN_
 from pyanaconda.payload import PackagePayload, payloadMgr, NoSuchGroup, PayloadError
-from pyanaconda.threads import threadMgr, AnacondaThread
+from pyanaconda.threading import threadMgr, AnacondaThread
 from pyanaconda import constants, iutil
 
 from pyanaconda.ui.communication import hubQ
