+++ pyanaconda/ui/tui/__init__.py
@@ -20,7 +20,7 @@
 from pyanaconda import ui
 from pyanaconda.ui.communication import hubQ
 from pyanaconda.flags import flags
-from pyanaconda.threads import threadMgr
+from pyanaconda.threading import threadMgr
 from pyanaconda.ui.tui import simpleline as tui
 from pyanaconda.ui.tui.hubs.summary import SummaryHub
 from pyanaconda.ui.tui.spokes import StandaloneSpoke
