+++ pyanaconda/ui/tui/spokes/progress.py
@@ -111,7 +111,7 @@ class ProgressSpoke(StandaloneTUISpoke):
 
     def refresh(self, args=None):
         from pyanaconda.install import doInstall, doConfiguration
-        from pyanaconda.threads import threadMgr, AnacondaThread
+        from pyanaconda.threading import threadMgr, AnacondaThread
 
         # We print this here because we don't really use the window object
         print(_(self.title))
